+++ tests/test_utils.py
@@ -1,0 +1,27 @@
+# StepUp Queue integrates queued jobs into a StepUp workflow.
+# © 2025 Toon Verstraelen
+#
+# This file is part of StepUp Queue.
+#
+# StepUp Queue is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 3
+# of the License, or (at your option) any later version.
+#
+# StepUp Queue is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, see <http://www.gnu.org/licenses/>
+#
+# --
+"""Unit tests for stepup.queue.utils."""
+
+from stepup.queue.utils import parse_sbatch
+
+
+def test_parse_sbatch():
+    assert parse_sbatch("123") == (123, None)
+    assert parse_sbatch("123;clu") == (123, "clu")